+++ packages/core/src/web/index.ts
@@ -3,3 +3,4 @@ export * from "./openapi";
 export * from "./playCount";
 export * from "./render";
 export * from "./scalar";
+export * from "./swapUrl";